PAR-332 Доработка отчёта Rep33B. Возврат строки подключения на прод.. Откат запроса данных из БД на decimal - решено брать округлённые значения, как в предыдущей версии отчёта.

+++ Rep33.Data/ReportData.cs
@@ -220,7 +220,7 @@ namespace Rep33.Data
             {
                 DataRow[] rs = _QueriesResult["SavedData"].Select($"VALUENAME  = '{ValDataName}'");
                 if (rs.Count<DataRow>() != 0)
-                    return rs[0].IsNull("VALUE") ? 0 : (decimal)rs[0].Field<double>("VALUE");
+                    return rs[0].IsNull("VALUE") ? 0 : rs[0].Field<decimal>("VALUE");
                 else return 0;
             }
         }
